nuplayer: Fix the "Fix CTS issues in NuPlayer"
If I play music with Google Play Music and then pause it, I have
two sources that are still polling for buffering updates in
the background.  This shows up in the log as

info/warning: (702,0)

Digging into it, the quoted change causes getCachedDuration to
never return the end of stream status (it always returns 0)
instead.

Fix this by not short-cutting sources that can do their own
cached lookup and instead move the new code down further in
the logic.

Change-Id: I7382561e2b3127016eb72307ed3c7f4e1fecc7b4